fix: changed quantity type to integer for pivot table order_product

+++ database/migrations/2025_12_23_010308_create_order_product.php
@@ -11,7 +11,7 @@ return new class extends Migration {
             $table->foreignId('order_id')->constrained('orders');
             $table->foreignId('product_id')->constrained('products');
             $table->decimal('price');
-            $table->decimal('quantity');
+            $table->integer('quantity');
         });
     }
